refactor: reintroduce profile update functionality with improved error handling and success messaging
